+++ requeriments.txt
@@ -1,0 +1,7 @@
+click
+google-api-python-client
+google-auth-httplib2
+google-auth-oauthlib
+imap-tools
+Jinja2
+tabulate
